Option to correct starting position
Previously the starting position in Othello was inverted, making the
engine incompatible with other engines.

The default has been left with the inverted position, for backwards
compatibility, but a new option --std-othello activates the correct
starting position.

Thanks to symmetries, the trained neural networks can indifferently
play in both cases.