First move is not searched but random
There is no point in spending resources launching a think()
procedure, as the first move is fixed apart for 4 symmetric
choices.

We prefer to choose the first move immediately, randomly, without
thinking.

As a consequence, there will be no more initial position in the
training data, but if the engine does not *think* at the first
move, it does not need the neural network to give correct
estimates at the starting position.

Moreover, the call to the network at the initial position is
meaningless also from a theoretical point of view: the policy is
useless, as there is only one legal move (with 4 symmetries); the
value is useless, as it should be the complementary of the value
at the second position, and that can be learned.